fix: now the notices component will await all request be completed to render their elements
