+++ src/components/organisms/Notices/components/listItem.tsx
@@ -2,10 +2,7 @@ import KeyboardArrowRightIcon from '@mui/icons-material/KeyboardArrowRight';
 import { StyledBox, StyledBtn, StyledTypo } from '../styles';
 import { NoticePreviewList } from '../types';
 import useNoticesService from '../services';
-import { useEffect, useState } from 'react';
-import { noticeSlugServices } from '@/app/home/notices/[id]/services';
 import { useSnackbar } from '@/context/snackbar-context';
-import { CulturalAgentPf, CulturalAgentPj } from '@/types';
 
 export default function NoticesListItem({
     id,
@@ -14,32 +11,12 @@ export default function NoticesListItem({
     endDate,
     router,
     setSelectedModule,
+    userPF,
+    userPJ,
 }: Readonly<NoticePreviewList>) {
     const { formatISODate, isDateGreaterThanToday } = useNoticesService;
-    const { getUserPFandPJ } = noticeSlugServices;
     const available = isDateGreaterThanToday(endDate);
     const { setSnackbar } = useSnackbar();
-    const [userPF, setUserPF] = useState<CulturalAgentPf | undefined>();
-    const [userPJ, setUserPJ] = useState<CulturalAgentPj | undefined>();
-
-    useEffect(() => {
-        const sessionJSON = localStorage.getItem('session');
-        if (!sessionJSON) {
-            setSnackbar({
-                message: 'token expirado, faça login novamente',
-                open: true,
-                severity: 'warning',
-            });
-            return router.push('/');
-        }
-        const { session } = JSON.parse(sessionJSON);
-
-        getUserPFandPJ(session.token).then((usersRegistrations) => {
-            setUserPF(usersRegistrations[0]);
-            setUserPJ(usersRegistrations[1]);
-        });
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
 
     function redirect() {
         if (!userPF && !userPJ) {
